centralize articles, video, and external data into global 11ty data
