+++ .prettierignore
@@ -1,6 +1,6 @@
 node_modules/
 package-lock.json
-/packages/lit-dev-content/site/
+/packages/lit-dev-content/site/tutorials/
 /packages/lit-dev-content/_site/
 /packages/*/lib/
 /packages/lit-dev-server/src/redirects.ts
