+++ packages/lit-dev-content/site/_data/externalSearchData.json
@@ -1,0 +1,13 @@
+[
+  {
+    "relativeUrl": "https://developer.mozilla.org/en-US/docs/Web/API/ElementInternals",
+    "title": "ElementInternals",
+    "heading": "",
+    "text": "The ElementInternals interface of the Document Object Model gives web developers a way to allow custom elements to fully participate in HTML forms. It provides utilities for working with these elements in the same way you would work with any standard HTML form element, and also exposes the Accessibility Object Model to the element.",
+    "docType": {
+      "type": "MDN",
+      "tag": "other"
+    },
+    "isExternal": true
+  }
+]
