+++ packages/lit-dev-content/site/articles/article/article.11tydata.js
@@ -1,6 +1,7 @@
 module.exports = {
   eleventyComputed: {
-    layout: "article",
-    permalink: data => `/articles/${data.page.fileSlug}/`,
-  }
-}
+    layout: 'article',
+    permalink: (data) => `/articles/${data.page.fileSlug}/`,
+  },
+  tags: ['articles'],
+};
